requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26
GitPython>=3.1.44
openai>=1.12.0